+++ packages/ratings/src/modules/review/models/review-activity.ts
@@ -5,6 +5,8 @@ export enum ReviewActivityType {
 	SUBMIT = 'submit',
 	APPROVE = 'approve',
 	REJECT = 'reject',
+	FEATURE = 'feature',
+	UNFEATURE = 'unfeature',
 	NOTE = 'note'
 }